+++ docs/README.md
@@ -1,0 +1,6 @@
+```Bash
+export PYTHONPATH=/home/src:$PYTHONPATH
+```
+
+To - Do: 
+    - [] check why output is generating to much noise
