+++ docs/repo_organization.md
@@ -1,0 +1,2 @@
+This could be of help, to understand more about how to organize the repo
+[Guide to python](https://docs.python-guide.org/writing/structure/)
