+++ main.py
@@ -1,4 +1,4 @@
-from DataGetter.DataArray import create_data_array
+from src.utilities.data_getter import create_data_array
 
 if __name__ == "__main__":
-    create_data_array("")
+    pass # for the moment please refer to ./test/test_data_getter.py
