--- optimization/DS/Objects/route.py
@@ -1,6 +1,0 @@
-
-
-class Route:
-    def __init__(number: str, directions: str, stops: strArr) -> None:
-        #!Requires implementation of an array of strings
-        pass
