+++ src/__init__.py
@@ -1,0 +1,3 @@
+""" Used to make logs of the modules"""
+import logging
+logging.basicConfig(level=logging.INFO)
