+++ src/utilities/data_getter/__init__.py
@@ -1,0 +1,1 @@
+from .data_array import create_data_array
