+++ src/utilities/data_structures/__init__.py
@@ -1,0 +1,3 @@
+from .bool_array import BoolArr
+from .int_array import IntArr
+from .obj_array import ObjArr
